+++ spec/registry_spec.rb
@@ -23,6 +23,13 @@ describe "openstack-image::registry" do
 
     expect_installs_curl
 
+    it "converges when configured to use sqlite" do
+      chef_run = ::ChefSpec::ChefRunner.new ::UBUNTU_OPTS
+      node = chef_run.node
+      node.set["openstack"]["db"]["image"]["db_type"] = "sqlite"
+      chef_run.converge "openstack-image::registry"
+    end
+
     it "installs mysql python packages" do
       expect(@chef_run).to install_package "python-mysqldb"
     end